backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pymysql>=1.1.0
python-multipart>=0.0.6
pyyaml>=6.0.0
python-dateutil>=2.8.0
passlib[bcrypt]>=1.7.4
itsdangerous>=2.1.0
requests>=2.31.0
pyjwt>=2.8.0
celery[redis]>=5.3.0
redis>=5.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
